Finish (#3)
Finish the compiler

+++ token.go
@@ -41,7 +41,7 @@ func NewStr(s string, pos Pos) *Token {
 	}
 }
 
-func NewInt(i int64, pos Pos) *Token {
+func NewInt(i int32, pos Pos) *Token {
 	return &Token{
 		tok:   "int",
 		value: i,
